+++ dependencyLibs/Core/src/org/droidplanner/core/drone/variables/MissionStats.java
@@ -6,7 +6,7 @@ import org.droidplanner.core.model.Drone;
 
 public class MissionStats extends DroneVariable {
     private double distanceToWp = 0;
-    private short currentWP = -1;
+    private int currentWP = -1;
 
     public MissionStats(Drone myDrone) {
         super(myDrone);
@@ -16,7 +16,7 @@ public class MissionStats extends DroneVariable {
         this.distanceToWp = disttowp;
     }
 
-    public void setWpno(short seq) {
+    public void setWpno(int seq) {
         if (seq != currentWP) {
             this.currentWP = seq;
             myDrone.notifyDroneEvent(DroneEventsType.MISSION_WP_UPDATE);
